init fix data race again

+++ ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: mamazzal <mamazzal@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/11 12:13:00 by mamazzal          #+#    #+#             */
-/*   Updated: 2023/05/30 16:27:04 by mamazzal         ###   ########.fr       */
+/*   Updated: 2023/06/01 14:27:36 by mamazzal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
-	int	ci;
-	int	fr;
-	int	isminus;
+	int			ci;
+	long long	fr;
+	int			isminus;
 
 	ci = 0;
 	isminus = 1;
